Module aws_sdk_honeycode::model

Modules

cell

See Cell

cell_input

See CellInput

column_metadata

See ColumnMetadata

create_row_data

See CreateRowData

data_item

See DataItem

delimited_text_import_options

See DelimitedTextImportOptions

destination_options

See DestinationOptions

failed_batch_item

See FailedBatchItem

filter

See Filter

import_data_source

See ImportDataSource

import_data_source_config

See ImportDataSourceConfig

import_job_submitter

See ImportJobSubmitter

import_options

See ImportOptions

result_row

See ResultRow

result_set

See ResultSet

source_data_column_properties

See SourceDataColumnProperties

table

See Table

table_column

See TableColumn

table_data_import_job_metadata

See TableDataImportJobMetadata

table_row

See TableRow

update_row_data

See UpdateRowData

upsert_row_data

See UpsertRowData

upsert_rows_result

See UpsertRowsResult

variable_value

See VariableValue

Structs

Cell

An object that represents a single cell in a table.

CellInput

ColumnMetadata

Metadata for column in the table.

CreateRowData

DataItem

The data in a particular data cell defined on the screen.

DelimitedTextImportOptions

DestinationOptions

An object that contains the options relating to the destination of the import request.

FailedBatchItem

Filter

ImportDataSource

An object that has details about the source of the data that was submitted for import.

ImportDataSourceConfig

ImportJobSubmitter

An object that contains the attributes of the submitter of the import job.

ImportOptions

An object that contains the options specified by the sumitter of the import request.

ResultRow

A single row in the ResultSet.

ResultSet

SourceDataColumnProperties

An object that contains the properties for importing data to a specific column in a table.

Table

An object representing the properties of a table in a workbook.

TableColumn

An object that contains attributes about a single column in a table

TableDataImportJobMetadata

The metadata associated with the table data import job that was submitted.

TableRow

An object that contains attributes about a single row in a table

UpdateRowData

UpsertRowData

UpsertRowsResult

VariableValue

The input variables to the app to be used by the InvokeScreenAutomation action request.

Enums

Format

ImportDataCharacterEncoding

ImportSourceDataFormat

TableDataImportJobStatus

UpsertAction